Move toolbar to layouts package
