+++ src/layouts/Toolbar.tsx
@@ -8,42 +8,43 @@
 
 import React from 'react';
 
-import Button from './components/Button';
-import { Toggle } from './hooks/useToggle';
-import DownArrowIcon from './icons/DownArrowIcon';
-import DownloadIcon from './icons/DownloadIcon';
-import FullScreenIcon from './icons/FullScreenIcon';
-import HandToolIcon from './icons/HandToolIcon';
-import HorizontalScrollingIcon from './icons/HorizontalScrollingIcon';
-import InfoIcon from './icons/InfoIcon';
-import LeftSidebarIcon from './icons/LeftSidebarIcon';
-import NextIcon from './icons/NextIcon';
-import PreviousIcon from './icons/PreviousIcon';
-import PrintIcon from './icons/PrintIcon';
-import RotateBackwardIcon from './icons/RotateBackwardIcon';
-import RotateForwardIcon from './icons/RotateForwardIcon';
-import TextSelectionIcon from './icons/TextSelectionIcon';
-import UpArrowIcon from './icons/UpArrowIcon';
-import VerticalScrollingIcon from './icons/VerticalScrollingIcon';
-import WrappedScrollingIcon from './icons/WrappedScrollingIcon';
-import ZoomInIcon from './icons/ZoomInIcon';
-import ZoomOutIcon from './icons/ZoomOutIcon';
-import { RenderToolbarSlot } from './layouts/ToolbarSlot';
-import LocalizationContext from './localization/LocalizationContext';
-import MoreActionsPopover, { ScrollMode } from './MoreActionsPopover';
-import OpenFileButton from './open/OpenFileButton';
-import PdfJs from './PdfJs';
-import Modal from './portal/Modal';
-import Position from './portal/Position';
-import Tooltip from './portal/Tooltip';
-import PropertiesModal from './property/PropertiesModal';
-import Match from './search/Match';
-import SearchPopover from './search/SearchPopover';
-import SelectionMode from './SelectionMode';
-import ThemeContent from './theme/ThemeContext';
+import Button from '../components/Button';
+import { Toggle } from '../hooks/useToggle';
+import DownArrowIcon from '../icons/DownArrowIcon';
+import DownloadIcon from '../icons/DownloadIcon';
+import FullScreenIcon from '../icons/FullScreenIcon';
+import HandToolIcon from '../icons/HandToolIcon';
+import HorizontalScrollingIcon from '../icons/HorizontalScrollingIcon';
+import InfoIcon from '../icons/InfoIcon';
+import LeftSidebarIcon from '../icons/LeftSidebarIcon';
+import NextIcon from '../icons/NextIcon';
+import PreviousIcon from '../icons/PreviousIcon';
+import PrintIcon from '../icons/PrintIcon';
+import RotateBackwardIcon from '../icons/RotateBackwardIcon';
+import RotateForwardIcon from '../icons/RotateForwardIcon';
+import TextSelectionIcon from '../icons/TextSelectionIcon';
+import UpArrowIcon from '../icons/UpArrowIcon';
+import VerticalScrollingIcon from '../icons/VerticalScrollingIcon';
+import WrappedScrollingIcon from '../icons/WrappedScrollingIcon';
+import ZoomInIcon from '../icons/ZoomInIcon';
+import ZoomOutIcon from '../icons/ZoomOutIcon';
+import { RenderToolbarSlot } from './ToolbarSlot';
+import LocalizationContext from '../localization/LocalizationContext';
+import OpenFileButton from '../open/OpenFileButton';
+import PdfJs from '../PdfJs';
+import Modal from '../portal/Modal';
+import Position from '../portal/Position';
+import Tooltip from '../portal/Tooltip';
+import PropertiesModal from '../property/PropertiesModal';
+import Match from '../search/Match';
+import SearchPopover from '../search/SearchPopover';
+import ScrollMode from '../ScrollMode';
+import SelectionMode from '../SelectionMode';
+import ThemeContent from '../theme/ThemeContext';
+import { decrease, increase, SpecialLevel } from '../zoom/zoomingLevel';
+import ZoomPopover from '../zoom/ZoomPopover';
+import MoreActionsPopover from './MoreActionsPopover';
 import './toolbar.less';
-import { decrease, increase, SpecialLevel } from './zoom/zoomingLevel';
-import ZoomPopover from './zoom/ZoomPopover';
 
 interface ToolbarProps {
     currentPage: number;
